Fix inspections + split classes into separate files
