+++ amf-aml/shared/src/main/scala/amf/client/model/domain/PropertyMapping.scala
@@ -23,68 +23,68 @@ case class PropertyMapping(override private[amf] val _internal: InternalProperty
   @JSExportTopLevel("model.domain.PropertyMapping")
   def this() = this(InternalPropertyMapping())
 
-  def withName(name: String) = {
+  def withName(name: String): PropertyMapping = {
     _internal.withName(name)
     this
   }
   def name(): StrField = _internal.name()
-  def withNodePropertyMapping(propertyId: String) = {
+  def withNodePropertyMapping(propertyId: String): PropertyMapping = {
     _internal.withNodePropertyMapping(propertyId)
     this
   }
   def nodePropertyMapping(): StrField = _internal.nodePropertyMapping()
-  def withLiteralRange(range: String) = {
+  def withLiteralRange(range: String): PropertyMapping = {
     _internal.withLiteralRange(range)
     this
   }
   def literalRange(): StrField = _internal.literalRange()
-  def withObjectRange(range: ClientList[String]) = {
+  def withObjectRange(range: ClientList[String]): PropertyMapping = {
     _internal.withObjectRange(range.asInternal)
     this
   }
   def objectRange(): ClientList[StrField] = _internal.objectRange().asClient
   def mapKeyProperty(): StrField          = _internal.mapKeyProperty()
-  def withMapKeyProperty(key: String) = {
+  def withMapKeyProperty(key: String): PropertyMapping = {
     _internal.withMapKeyProperty(key)
     this
   }
   def mapValueProperty(): StrField = _internal.mapKeyProperty()
-  def withMapValueProperty(value: String) = {
+  def withMapValueProperty(value: String): PropertyMapping = {
     _internal.withMapValueProperty(value)
     this
   }
   def minCount(): IntField = _internal.minCount()
-  def withMinCount(minCount: Int) = {
+  def withMinCount(minCount: Int): PropertyMapping = {
     _internal.withMinCount(minCount)
     this
   }
   def pattern(): StrField = _internal.pattern()
-  def withPattern(pattern: String) = {
+  def withPattern(pattern: String): PropertyMapping = {
     _internal.withPattern(pattern)
     this
   }
   def minimum(): DoubleField = _internal.minimum()
-  def withMinimum(min: Double) = {
+  def withMinimum(min: Double): PropertyMapping = {
     _internal.withMinimum(min)
     this
   }
   def maximum(): DoubleField = _internal.maximum()
-  def withMaximum(max: Double) = {
+  def withMaximum(max: Double): PropertyMapping = {
     _internal.withMaximum(max)
     this
   }
   def allowMultiple(): BoolField = _internal.allowMultiple()
-  def withAllowMultiple(allow: Boolean) = {
+  def withAllowMultiple(allow: Boolean): PropertyMapping = {
     _internal.withAllowMultiple(allow)
     this
   }
   def enum(): ClientList[AnyField] = _internal.enum().asClient
-  def withEnum(values: ClientList[Any]) = {
+  def withEnum(values: ClientList[Any]): PropertyMapping = {
     _internal.withEnum(values.asInternal)
     this
   }
   def sorted(): BoolField = _internal.sorted()
-  def withSorted(sorted: Boolean) = {
+  def withSorted(sorted: Boolean): PropertyMapping = {
     _internal.withSorted(sorted)
     this
   }
@@ -101,13 +101,13 @@ case class PropertyMapping(override private[amf] val _internal: InternalProperty
 
   def withTypeDiscriminator(typesMapping: ClientMap[String]) = throw new Exception("Not implemented yet")
   def typeDiscriminatorName(): StrField                      = _internal.typeDiscriminatorName()
-  def withTypeDiscriminatorName(name: String) = {
+  def withTypeDiscriminatorName(name: String): PropertyMapping = {
     _internal.withTypeDiscriminatorName(name)
     this
   }
 
-  def withExternallyLinkable(linkable: Boolean) = _internal.withExternallyLinkable(linkable);
-  def externallyLinkable(): BoolField           = _internal.externallyLinkable()
+  def withExternallyLinkable(linkable: Boolean): InternalPropertyMapping = _internal.withExternallyLinkable(linkable);
+  def externallyLinkable(): BoolField                                    = _internal.externallyLinkable()
 
   def classification(): String = {
     _internal.classification() match {
