+++ amf-aml/shared/src/main/scala/amf/plugins/document/vocabularies/metamodel/document/DialectModel.scala
@@ -11,18 +11,19 @@ import amf.plugins.document.vocabularies.model.document.Dialect
 
 object DialectModel extends DocumentModel with ExternalContextModel {
 
-  val Name =
+  val Name: Field =
     Field(Str, Namespace.Core + "name", ModelDoc(ModelVocabularies.Core, "name", "Name of the dialect"))
-  val Version =
+
+  val Version: Field =
     Field(Str, Namespace.Core + "version", ModelDoc(ModelVocabularies.Core, "version", "Version of the dialect"))
-  val Documents = Field(DocumentsModelModel,
-                        Namespace.Meta + "documents",
-                        ModelDoc(ModelVocabularies.Meta, "documents", "Document mapping for the the dialect"))
+
+  val Documents: Field = Field(DocumentsModelModel,
+                               Namespace.Meta + "documents",
+                               ModelDoc(ModelVocabularies.Meta, "documents", "Document mapping for the the dialect"))
 
   override def modelInstance: AmfObject = Dialect()
 
-  override val `type`: List[ValueType] =
-    Namespace.Meta + "Dialect" :: DocumentModel.`type`
+  override val `type`: List[ValueType] = Namespace.Meta + "Dialect" :: DocumentModel.`type`
 
   override val fields
     : List[Field] = Name :: Version :: Externals :: Documents :: BaseUnitModel.Location :: DocumentModel.fields
