+++ amf-aml/shared/src/main/scala/amf/plugins/document/vocabularies/metamodel/domain/ExternalModel.scala
@@ -9,10 +9,10 @@ import amf.plugins.document.vocabularies.model.domain.External
 
 object ExternalModel extends DomainElementModel {
 
-  val DisplayName = Field(Str,
-                          Namespace.Core + "displayName",
-                          ModelDoc(ModelVocabularies.Core, "displayName", "The display name of the item"))
-  val Base =
+  val DisplayName: Field = Field(Str,
+                                 Namespace.Core + "displayName",
+                                 ModelDoc(ModelVocabularies.Core, "displayName", "The display name of the item"))
+  val Base: Field =
     Field(Str, Namespace.Meta + "base", ModelDoc(ModelVocabularies.Meta, "base", "Base URI for the external model"))
 
   override def modelInstance: AmfObject = External()
