+++ amf-aml/shared/src/main/scala/amf/plugins/document/vocabularies/metamodel/domain/VocabularyReferenceModel.scala
@@ -8,9 +8,11 @@ import amf.core.vocabulary.{Namespace, ValueType}
 import amf.plugins.document.vocabularies.model.domain.VocabularyReference
 
 object VocabularyReferenceModel extends DomainElementModel {
-  val Alias     = Field(Str, Namespace.Document + "alias")
-  val Reference = Field(Str, Namespace.Document + "reference")
-  val Base      = Field(Str, Namespace.Meta + "base")
+  val Alias: Field = Field(Str, Namespace.Document + "alias")
+
+  val Reference: Field = Field(Str, Namespace.Document + "reference")
+
+  val Base: Field = Field(Str, Namespace.Meta + "base")
 
   override def modelInstance: AmfObject = VocabularyReference()
